Fix travis

+++ dreamchess/src/gui/xmlsavefile.cpp
@@ -65,7 +65,7 @@ void write_save_xml( int slot )
     pugi::xml_node save = doc.append_child("save");
 
     pugi::xml_node node = save.append_child("time");
-    node.text().set(std::chrono::system_clock::to_time_t(std::chrono::system_clock::now()));
+    node.text().set(static_cast<long long>(std::chrono::system_clock::to_time_t(std::chrono::system_clock::now())));
 
     if (ch_userdir()) {
         DBG_WARN("could not enter user directory");
